% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportButton.R
\name{reportButton}
\alias{reportButton}
\title{Report download button}
\usage{
reportButton(inputId, label = "View Report", icon = NULL, ...)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{The contents of the button or link--usually a text label, but
you could also use any other HTML, like an image.}

\item{icon}{An optional \code{\link[shiny:icon]{icon()}} to appear on the button.}

\item{...}{Named attributes to be applied to the button or link.}
}
\value{
A \code{shiny.tag.list} object definition (from
\code{\link[shiny]{tagList}}) containing two components:
\itemize{
\item A \code{\link[shiny]{singleton}} \code{<script>} tag, injected
into the page \code{<head>}, that registers a custom 'shiny' message
handler (\code{"openWindow"}) for opening the rendered report in a
new browser tab/window.
\item A \code{\link[shiny]{actionButton}} with the argument definitions
of \code{inputId}, \code{label} and \code{icon}.
}
The return value should only be included under a
\code{\link[shiny]{fluidPage}} 'shiny' definition.
}
\description{
Modifies \code{\link[shiny:actionButton]{shiny::actionButton()}} to generate and open an R
Markdown report in a new window tab.
}
\examples{
if (interactive()) {
  ui <- fluidPage(
    reportButton(
      inputId = "knit_report",
      label = "Generate Report"
    )
  )
}
}
\seealso{
\code{\link[=renderReport]{renderReport()}}
}
