% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderReport.R
\name{renderReport}
\alias{renderReport}
\title{Render report to a new window}
\usage{
renderReport(
  inputId,
  title = "Report",
  rmd_file,
  params = NULL,
  envir = parent.frame()
)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{title}{The title to label the report tab.}

\item{rmd_file}{The input file to be rendered (.Rmd).}

\item{params}{A list of named parameters that override custom params specified within the YAML front-matter.}

\item{envir}{The environment in which the code chunks are to be evaluated during knitting.}
}
\description{
Server-side logic to compile .Rmd and send HTML to client.
}
\examples{
\dontrun{
server <- function(input, output, session) {
   renderReport(
     inputId = "knit_report",
     title = "My Report",
     rmd_file = "report.Rmd",
     params = list(date = Sys.Date())
   )
}
}
}
\seealso{
\code{\link[=reportButton]{reportButton()}}
}
