% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportButton.R
\name{reportButton}
\alias{reportButton}
\title{Report download button}
\usage{
reportButton(inputId, label = "View Report", icon = NULL, ...)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{The contents of the button or link--usually a text label, but
you could also use any other HTML, like an image.}

\item{icon}{An optional \code{\link[shiny:icon]{icon()}} to appear on the button.}

\item{...}{Named attributes to be applied to the button or link.}
}
\description{
Modifies \code{\link[shiny:actionButton]{shiny::actionButton()}} to generate and open an R Markdown report in a new window tab.
}
\seealso{
\code{\link[=renderReport]{renderReport()}}
}
